==> Songs_Collections_Cleaning/1-Collecting_Top_Songs.ipynb <==
## Import libraries

import autoreload
%load_ext autoreload
%autoreload 2
%reload_ext autoreload
from config import username, password
from datetime import timedelta, date
import pandas as pd
import Classes
import os
import config
import time
import random

# Setup chromedriver
from selenium import webdriver
from selenium.webdriver.common.keys import Keys
from selenium.webdriver.common.by import By

## Collect top daily Tiktok Daily Songs 
- I paid for the Chartmetric subscription given they have the most extensive (and really only) database of TikTok charts                         

#### Chartmetrics daily charts can be accessed using a particular combination of a url + date format

# Creat list of urls for daily tiktok charts

# End date for daily urls function
end_date = date(2021,3,8)

# Contains all links to each daily chart
lst_daily_urls = Classes.daily_urls(end_date)

# Specify the path wheere I want all csvs to be downloaded

options = webdriver.ChromeOptions()

preferences = {"download.default_directory":"C:\\Users\\Andrew\\Documents\\Metis\\TikTok_Hit_Predictor\\Songs_Data"}

# Add preferences to options
options.add_experimental_option("prefs",preferences)

# Create executable path for chromedriver
chromedriver = '/Program Files/Google/Chrome/Application/chromedriver.exe' # path to the chromedriver executable
os.environ["webdriver.chrome.driver"] = chromedriver

# Define email and password xpaths
email_xpath = '//*[@id="email"]'

password_xpath = '//*[@id="password"]'

login_xpath = '/html/body/div[1]/div[2]/login/div/section/div/form/div[4]/button'

# Create driver
driver = webdriver.Chrome(chromedriver,options=options)

for index, url in enumerate(lst_daily_urls):
    if index == 0:
        driver.get(url)
        driver.refresh()
        time.sleep(5+5*random.random())
        # Input username and password
        driver.find_element(By.XPATH,email_xpath).send_keys(config.username)
        driver.find_element(By.XPATH,password_xpath).send_keys(config.password)
        driver.find_element(By.XPATH,login_xpath).click()
        time.sleep(2+5*random.random())
        driver.get(url)
        driver.refresh()
        time.sleep(20+5*random.random())
        driver.find_element(By.XPATH,"/html/body/div[1]/div[2]/div/div[3]/ui-view/div[2]/div/div/div[1]/div/export-csv-button").click()
    else:
        try: 
            driver.get(url)
            driver.refresh()
            time.sleep(5+5*random.random())
            driver.find_element(By.XPATH,"/html/body/div[1]/div[2]/div/div[3]/ui-view/div[2]/div/div/div[1]/div/export-csv-button").click() 
        except:
            driver.get(url)
            driver.refresh()
            time.sleep(15+5*random.random())
            driver.find_element(By.XPATH,"/html/body/div[1]/div[2]/div/div[3]/ui-view/div[2]/div/div/div[1]/div/export-csv-button").click()            


## Collect Weekly Top TikTok Songs
There is a database for 7 day rolling top songs.  To ensure we have a robust dataset, I also download all of these csvs.

# Create a list of urls 

# End date for 7 day rolling urls function
end_date = date(2021,3,9)

# Contains all links to each 7 day rolling chart
lst_weekly_rolling_urls = Classes.weekly_rolling_urls(end_date)

# Specify the path wheere I want all csvs to be downloaded

options = webdriver.ChromeOptions()

preferences = {"download.default_directory":"C:\\Users\\Andrew\\Documents\\Metis\\TikTok_Hit_Predictor\\Songs_Data\\Rolling_Weekly"}

# Add preferences to options
options.add_experimental_option("prefs",preferences)

# Create executable path for chromedriver
chromedriver = '/Program Files/Google/Chrome/Application/chromedriver.exe' # path to the chromedriver executable
os.environ["webdriver.chrome.driver"] = chromedriver

# Define email and password xpaths
email_xpath = '//*[@id="email"]'

password_xpath = '//*[@id="password"]'

login_xpath = '/html/body/div[1]/div[2]/login/div/section/div/form/div[4]/button'

csv_xpath = '/html/body/div[1]/div[2]/div/div[3]/ui-view/div[2]/div/div/div[1]/div/export-csv-button/button'

# Create driver
driver = webdriver.Chrome(chromedriver,options=options)

for index, url in enumerate(lst_weekly_rolling_urls[24:]):
    if index == 0:
        driver.get(url)
        driver.refresh()
        time.sleep(5+5*random.random())
        # Input username and password
        driver.find_element(By.XPATH,email_xpath).send_keys(config.username)
        driver.find_element(By.XPATH,password_xpath).send_keys(config.password)
        driver.find_element(By.XPATH,login_xpath).click()
        time.sleep(2+5*random.random())
        driver.get(url)
        driver.refresh()
        time.sleep(20+5*random.random())
        driver.find_element(By.XPATH,csv_xpath).click()
    else:
        try: 
            driver.get(url)
            driver.refresh()
            time.sleep(2+5*random.random())
            driver.find_element(By.XPATH,csv_xpath).click() 
        except:
            try:        
                driver.get(url)
                driver.refresh()
                time.sleep(15+5*random.random())
                driver.find_element(By.XPATH,csv_xpath).click()
            except:
                driver.get(url)
                driver.refresh()
                time.sleep(15+5*random.random())
                driver.find_element(By.XPATH,csv_xpath).click()
