==> Songs_Collections_Cleaning/2-Song_Cleaning.ipynb <==
# Notebook Summary
In this notebook I will concat all individual rolling weekly csvs into a single csv, all daily csvs into a single csv and then pull out all unique songs with their metadata.

import os
import glob
import pandas as pd
import pickle

# Starting with daily csvs

# Designate path

os.chdir('C:\\Users\\Andrew\\Documents\\Metis\\TikTok_Hit_Predictor\\Songs_Data\\Daily')

# Create a list of all filenames 

extension = 'csv'
all_filenames = [i for i in glob.glob('*.{}'.format(extension))]

# Combine all files in the list

daily_combined_csv = pd.concat([pd.read_csv(f) for f in all_filenames])

#export to csv
daily_combined_csv.to_csv( "combined_csv.csv", index=False, encoding='utf-8-sig')

# Next daily top songs calculated based on a 7 day moving average

# Designate path

os.chdir('C:\\Users\\Andrew\\Documents\\Metis\\TikTok_Song_Predictor\\Songs_Data\\Rolling_Weekly')

# Create a list of all filenames 

extension = 'csv'
all_filenames = [i for i in glob.glob('*.{}'.format(extension))]

# Combine all files in the list

rolling_combined_csv = pd.concat([pd.read_csv(f) for f in all_filenames])

#export to csv
rolling_combined_csv.to_csv( "combined_csv.csv", index=False, encoding='utf-8-sig')

# Selecting out unique songs and combine into a single dataframe

#### Start with rolling 7 day daily ranking

# Create a new datafrane where each row is the maximum rank for each song

unique_combined_rolling = rolling_combined_csv.sort_values('Rank',ascending=True).drop_duplicates(['tiktok_track_id'])

unique_combined_rolling.shape

#### Next Daily ranking

# Create a new datafrane where each row is the maximum rank for each song

unique_combined_daily = daily_combined_csv.sort_values('Rank',ascending=True).drop_duplicates(['tiktok_track_id'])

unique_combined_daily.shape

#### Combine unique dataframes

# Check the columns for both

unique_combined_rolling.columns

unique_combined_daily.columns

# Let's drop the 'Weekly Posts' column before concating the two dataframes

unique_combined_rolling.drop(['Weekly Posts'],axis = 1,inplace=True)

# Concat both dataframes

agg_combined = pd.concat([unique_combined_daily,unique_combined_rolling])

# Delete duplicates by only selecting the min rank again

agg_combined = agg_combined.sort_values('Rank',ascending=True).drop_duplicates(['tiktok_track_id'])

#### Pickle agg_combined dataframe

# Designate path

path = r"C:\Users\Andrew\Documents\Metis\TikTok_Hit_Predictor\Pickle\Pickle.pkl"
with open(path, 'wb') as f:
    pickle.dump(agg_combined, f)
