==> Songs_Collections_Cleaning/4-Spotify_Feature_Extraction.ipynb <==
import autoreload
%load_ext autoreload
%autoreload 2
%reload_ext autoreload
import config

import requests
import base64
import datetime
from urllib.parse import urlencode
import pickle
import pandas as pd
import Spotify_Extraction

# Pickle dump song dataset and save into dataframe

# Load Pickled Dataframe from Cleaning Notebook into a DataFrame
path = r"C:\Users\Andrew\Documents\Metis\TikTok_Hit_Predictor\Pickle\Spotify_Pickle.pkl"

df = pickle.load(open(path,'rb'))

# Reset Index
df.reset_index(inplace=True)

# Extract Features

spotify = Spotify_Extraction.SpotifyAPI(config.client_id, config.client_secret)

# Function to extract song_id 

def extract_id(isrc):
    try: 
        song_id = spotify.search(f"isrc:{isrc}", search_type = "Track")['tracks']['items'][0]['id']
    except:
        song_id = {}
    return song_id

# Create an id column

df['spotify_uri'] = df['ISRC'].apply(lambda x : extract_id(x))

# Create audio analysis column
df['audio_analysis'] = df['spotify_uri'].apply(lambda x : spotify.get_audio_analysis(x))

# Create feature analysis column
df['feature_analysis'] = df['spotify_uri'].apply(lambda x : spotify.get_audio_features(x))

df.head()

# Pickle the cleaned dataframe

# Designate path

path = r"C:\Users\Andrew\Documents\Metis\TikTok_Hit_Predictor\Pickle\Feauture_Pickle.pkl"
with open(path, 'wb') as f:
    pickle.dump(df, f)
