==> Songs_Collections_Cleaning/5-Post_Spotify_Extraction_Cleaning.ipynb <==
# Notebook Summary
This notebook cleans the extracted Spotify data for analysis

import pickle
import pandas as pd
import numpy as np

# Pickle in spotify data

# Pickle in cleaned dataframe

# Designate path

path = r"C:\Users\Andrew\Documents\Metis\TikTok_Song_Predictor\Pickle\Feauture_Pickle.pkl"

df = pickle.load(open(path,'rb'))
df.head()

df.shape

# Extractng 'feature analysis' column into separate columns

list_features = [] # list to store each row in 

for index, row in df.iterrows():
    my_dict = df['feature_analysis'][index]
    exploded_features = pd.DataFrame.from_dict(my_dict,orient='index').T
    exploded_features['ISRC'] = df['ISRC'][index]
    # Concat with df
    list_features.append(exploded_features)

df_features = pd.concat(list_features)

df_features.shape

df_features.head()

df.shape

# Merge df_features with df

# Perfrom a left merge with df_features

df_agg = df_features.merge(df,how='inner',on='ISRC')

df_agg.columns

# Drop rows were tiktok id is not unique, keep the row were 

df_agg = df_agg.sort_values('Peak Position').drop_duplicates('uri', keep = 'first')

df_agg['success'] =  np.where(df_agg['Peak Position'] < 51, 1, 0)

# Further Data Cleaning

# check for distribution of the target
target_pcts = df_agg.success.value_counts(normalize=True) 
print(target_pcts)

# Add Further Features

# Change dtype

df_agg.dtypes

# Convert columns to float 64
cols = ['danceability', 'energy', 'key', 'loudness', 'mode', 'speechiness',
       'acousticness', 'instrumentalness', 'liveness', 'valence', 'tempo','duration_ms',
       'time_signature','Velocity']
df_agg[cols] = df_agg[cols].apply(pd.to_numeric, errors='coerce', axis=1)

df_agg['Release Date']

df_agg['Release Date']= pd.to_datetime(df['Release Date'],errors='coerce')

df_agg.dtypes

# Create a year column

df_agg['year'] = pd.DatetimeIndex(df_agg['Release Date']).year

# Reset Index
df_agg.reset_index(inplace=True)

df_agg['year'] = df_agg['year'].fillna(0)

## Top / "Other" Albums and Artists column to make getting dummies manageable

### Start with albums

# Adding a column for the top 10 artists and top ten albums (otherwise with get dummies -> too many columns)

album_counts = df_agg['Album Name'].value_counts()
df_album = pd.DataFrame(album_counts)

album_counts[0:30]

# Creating list of "other albums with less than 5 appearances"
other_albums = list(album_counts[album_counts <= 30].index)

df_agg['top_albums'] = df_agg['Album Name'].replace(other_albums,'Other')

### Next artists

# Adding a column for the top 10 artists and top ten albums (otherwise with get dummies -> too many columns)

artist_counts = df_agg['Artist(s)'].value_counts()
df_album = pd.DataFrame(artist_counts)

artist_counts[0:100]

# Creating list of "other albums with less than 5 appearances"
other_artists = list(artist_counts[artist_counts <= 7].index)

df_agg['top_artists'] = df_agg['Artist(s)'].replace(other_artists,'Other')

df_agg['top_artists'].value_counts()

# Pickle dataset for modelling

# Designate path

path = r"C:\Users\Andrew\Documents\Metis\TikTok_Song_Predictor\Pickle\df_agg.pkl"
with open(path, 'wb') as f:
    pickle.dump(df_agg, f)

df_agg.shape
