==> Songs_Collections_Cleaning/6-Outside_Data_Cleaning.ipynb <==
# Notebook Summary
In this notebook I clean the Kaggle dataset so that the XGB model can predictions can be made on it

import os
import pandas as pd
import pickle

# Import CSV

# Designate path

data = ('C:\\Users\\Andrew\\Documents\\Metis\\TikTok_Song_Predictor\\Songs_Data\\Spotify_Kaggle\\data.csv')

df = pd.read_csv(data)

df.head()

# Cleaning

# Search for columns with null or missing values

df[df.isnull().any(axis=1)]

No coluumns with null data

# inspect dtypes

df.dtypes

# Pickle Dump

# Designate path

path = r"C:\Users\Andrew\Documents\Metis\TikTok_Song_Predictor\Pickle\Kaggle.pkl"
with open(path, 'wb') as f:
    pickle.dump(df, f)
