==> Songs_Collections_Cleaning/7-Track_Artists.ipynb <==
# Notebook Summary
In this notebook, I create a list of Artists for each song on Spotify.  I cannot apply the Chartmetric artists names directly to the Spotify artist names from the outside dataset. 

import autoreload
%load_ext autoreload
%autoreload 2
%reload_ext autoreload
import config

import numpy as np
import requests
import base64
import datetime
from urllib.parse import urlencode
import pickle
import pandas as pd
import Spotify_Extraction

# Pickle in data

# Designate path

path = r"C:\Users\Andrew\Documents\Metis\TikTok_Hit_Predictor\Pickle\df_agg.pkl"

df = pickle.load(open(path,'rb'))
df.head()

# Extract Artist Name

spotify = Spotify_Extraction.SpotifyAPI(config.client_id, config.client_secret)

# Function to extract artists

def extract_artists(isrc):
    try: 
        song_dict = spotify.search(f"isrc:{isrc}", search_type = "Track")
        artists = []
        for i in song_dict.get('tracks').get('items')[0].get('artists'):
            artists.append(i.get('name'))
    except:
        artists = []
    return artists

# Create a spotify artists column

df['spotify_artists'] = df['ISRC'].apply(lambda x : extract_artists(x))

# Inspect
df.head()

# Convert from list to string

# Create helper function

# Python program to convert a list 
# to string using join() function 
    
# Function to convert   
def listToString(s):  
    
    # initialize an empty string 
    str1 = " " 
    
    # return string   
    return (str1.join(s))

df['spotify_artists'] = df['spotify_artists'].apply(lambda x: listToString(x))

df.head()

# Replace blanks strings with nan
df['spotify_artists'].replace('Other', np.nan, inplace=True)

df.shape

# Pickle Dump

path = r"C:\Users\Andrew\Documents\Metis\TikTok_Hit_Predictor\Pickle\df_agg_w_artists.pkl"

with open(path, 'wb') as f:
    pickle.dump(df, f)
