==> Songs_Collections_Cleaning/8-Categorical_Encoding.ipynb <==
# Notebook Summary
In this notebook, I combine the hit/not a hit dataset with the outside Kaggle dataset.  I am doign this so that the categorical encodings in both datasets is synonymous.  I will do this using two methods: dummy variables and factorizing.  I consider the data where we know whether the song is a success or not as "supervised" and where we do not know "unsupervised"

import pickle 
import pandas as pd
import numpy as np
import string
import Classes

# Pickle in supervised dataset

path = r"C:\Users\Andrew\Documents\Metis\TikTok_Hit_Predictor\Pickle\df_agg_w_artists.pkl"

df_supervised = pickle.load(open(path,'rb'))
df_supervised.head()

# Pickle in unsupervised dataset (predictions to be placed on this set)

# Pickle in cleaned dataframe

# Designate path

path = r"C:\Users\Andrew\Documents\Metis\TikTok_Song_Predictor\Pickle\Kaggle.pkl"

df_unsupervised = pickle.load(open(path,'rb'))
df_unsupervised.head(2)

### I need to remove the lists from the "unsupervised" dataset

df_unsupervised['artists']

df_supervised['spotify_artists'].tail(10)

# Function to remove brackets and quotes from unsupervised learning dataset

def delist(s):
    
    # remove brackets
    str1 = s[1:-2]
    
    # Replace quotes
    str1 = str1.replace("'",'')
    
    # Replace commas
    str1 = str1.replace(",",'')
    
    return str1

# Create a column called "spotify_artists" in this format

df_unsupervised['spotify_artists'] = df_unsupervised['artists'].apply(lambda x: delist(x))

# Inspect

df_unsupervised.head()

# Ensuring column names are the same in both dataframes

df_supervised.columns

df_unsupervised.columns

# Select specific features

X_supervised = df_supervised.loc[:,['danceability', 'energy', 'key', 'loudness', 'mode', 'speechiness',
       'acousticness', 'instrumentalness', 'liveness', 'valence', 'tempo','year','spotify_artists','success']]

# Select specific features

X_unsupervised = df_unsupervised.loc[:,['danceability', 'energy', 'key', 'loudness', 'mode', 'speechiness',
       'acousticness', 'instrumentalness', 'liveness', 'valence', 'tempo','year','spotify_artists']]

# Start with Factorizing
This method works well over dummy variables when dummy variables will result in an additional large number of dimensions, leads to curse of dimensionality

# Get lenght of the supervised dataset so that I can index it later on
supervised_len = len(X_supervised)

# Combine datasets
dataset_factorize = pd.concat(objs=[X_supervised, X_unsupervised], axis = 0)

# Factorize
Classes.gnumeric_func(dataset_factorize,['spotify_artists'])

### Split based on supervised vs. unserpvised

# Split based on supversied vs. unsupervised

supervised_factorized = dataset_factorize[:supervised_len]

# Check shape
supervised_factorized.shape

# Split based on supversied vs. unsupervised

unsupervised_factorized = dataset_factorize[supervised_len:]

# Check shape
unsupervised_factorized.shape

### Pickle

# Pickle dump these factorized datasets

# Start with supervised

path = r"C:\Users\Andrew\Documents\Metis\TikTok_Hit_Predictor\Pickle\supervised_factorized.pkl"

with open(path, 'wb') as f:
    pickle.dump(supervised_factorized, f)

# Pickle dump these factorized datasets

# Start with supervised

path = r"C:\Users\Andrew\Documents\Metis\TikTok_Hit_Predictor\Pickle\unsupervised_factorized.pkl"

with open(path, 'wb') as f:
    pickle.dump(unsupervised_factorized, f)

# Create Dummy Variables for top artists - "Other" group method

# Select specific features

X_supervised = df_supervised.loc[:,['danceability', 'energy', 'key', 'loudness', 'mode', 'speechiness',
       'acousticness', 'instrumentalness', 'liveness', 'valence', 'tempo','year','spotify_artists','success']]

# Select specific features

X_unsupervised = df_unsupervised.loc[:,['danceability', 'energy', 'key', 'loudness', 'mode', 'speechiness',
       'acousticness', 'instrumentalness', 'liveness', 'valence', 'tempo','year','spotify_artists']]

# Get lenght of the supervised dataset so that I can index it later on
supervised_len = len(X_supervised)

# Combine datasets
dataset_dummy = pd.concat(objs=[X_supervised, X_unsupervised], axis = 0)

# Create counts of all artists
artist_counts = dataset_dummy.spotify_artists.value_counts()

artist_counts[0:500]

# Define other artists
other_artist = list(artist_counts[artist_counts < 50].index)

# Replace these artists with 'other'
dataset_dummy['spotify_artists'] = dataset_dummy['spotify_artists'].replace(other_artist,'other')

# Inspect
dataset_dummy.head(200)

#get dummies
dataset_dummy = pd.get_dummies(dataset_dummy)

### Split based on supervised vs. unsupervised

# Split based on supversied vs. unsupervised

supervised_dummy = dataset_dummy[:supervised_len]

# Split based on supversied vs. unsupervised

unsupervised_dummy = dataset_dummy[supervised_len:]

supervised_dummy['success']

### Pickle

# Pickle dump these factorized datasets

# Start with supervised

path = r"C:\Users\Andrew\Documents\Metis\TikTok_Hit_Predictor\Pickle\supervised_dummy.pkl"

with open(path, 'wb') as f:
    pickle.dump(supervised_dummy, f)

# Pickle dump these factorized datasets

# Start with supervised

path = r"C:\Users\Andrew\Documents\Metis\TikTok_Hit_Predictor\Pickle\unsupervised_dummy.pkl"

with open(path, 'wb') as f:
    pickle.dump(unsupervised_dummy, f)
